+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Product;
 use App\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
 use DB;
 
 class ProductController extends Controller
@@ -16,11 +17,16 @@ class ProductController extends Controller
      */
     public function index()
     {
+        if(Session::get('login') && ((Session('type') == 4 || Session('type') == 2) || (Session('type') == 1))){
         $products = DB::table('product')
                 ->join('categories', 'categories.category_id', '=', 'product.category_id')
                 ->select('product.product_id','categories.category_name','product.product_name','product.product_price','product.product_stock','product.explanation')
                 ->get();
         return view('product/list', ['products' => $products]);
+        }
+        else{
+            return redirect('/')->with('alert','Anda tidak memiliki akses ke halaman');
+        }
     }
 
     /**
@@ -30,8 +36,13 @@ class ProductController extends Controller
      */
     public function create()
     {
+        if(Session::get('login') && (Session('type') == 4)){
         $categories = Category::all();
         return view('product/form', ['categories' => $categories]);
+    }
+        else{
+            return redirect('/')->with('alert','Anda tidak memiliki akses ke halaman');
+        }
     }
 
     /**
@@ -46,6 +57,7 @@ class ProductController extends Controller
                         'product_name' => 'required',
                         'product_price' => 'required',
                         'product_stock' => 'required']);
+        if(Session::get('login') && (Session('type') == 4)){
         $price = $request->input('product_price');
         $price = str_replace("Rp ","",$price);
         $price = str_replace(".","",$price);
@@ -57,17 +69,11 @@ class ProductController extends Controller
                         'explanation' => e($request->input('explanation')) ]);
         return redirect()->route('product.index')->with('inserted',$request->input('product_name'));
     }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Product  $product
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Product $product)
-    {
-        //
+        else{
+            return redirect('/')->with('alert','Anda tidak memiliki akses ke halaman');
+        }
     }
+
 
     /**
      * Show the form for editing the specified resource.
@@ -77,9 +83,14 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
+        if(Session::get('login') && (Session('type') == 4)){
         $categories = Category::all();
         $product = Product::find($id);
         return view('/product/edit', ['product' => $product, 'categories' => $categories]);
+    }
+        else{
+            return redirect('/')->with('alert','Anda tidak memiliki akses ke halaman');
+        }
     }
 
     /**
@@ -95,6 +106,7 @@ class ProductController extends Controller
                         'product_name' => 'required',
                         'product_price' => 'required',
                         'product_stock' => 'required']);
+        if(Session::get('login') && (Session('type') == 4)){
         $product = Product::find($id);
         $price = $request->input('product_price');
         $price = str_replace("Rp ","",$price);
@@ -107,6 +119,10 @@ class ProductController extends Controller
         $product->save();
         return redirect()->route('product.index')->with('edited',$request->input('product_name'));
     }
+        else{
+            return redirect('/')->with('alert','Anda tidak memiliki akses ke halaman');
+        }
+    }
 
     /**
      * Remove the specified resource from storage.
@@ -116,8 +132,13 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
+        if(Session::get('login') && (Session('type') == 1)){
         $product = Product::find($id);
         $product->delete();
         return redirect()->route('product.index')->with('deleted',$id);
+        }
+        else{
+            return redirect('/')->with('alert','Anda tidak memiliki akses ke halaman');
+        }
     }
 }
